Daily save
Notes: instability observed in fluid-solid interface when solid is very soft (pain-water interface in Ethan's work). Seems to be suppressible by reducing timestep, so probably this is real thing due to presence of damping matrix. Bigger issue is that the boundary node tracing is still not fully robust. OPG demo shows one with leakage out even in pristine domains. Almost certainly something to do with interface elements causing problems as this seems to be where it goes wrong. Need to get a more robust way of doing this - could we go from global matrix directly to necessary nodes?