Daily save
Fixed stupid bug in OPG test (it was in the order of operations when creating r-theta mesh - need to do fluid-solid interface last), but this also forced tidy up of subdomain model execution code, which is now much cleaner and more readable. Final thing to do is validation case code.